% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newlinreg1.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\value{
a value linreg.
}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{beta_hat}}{matrix.}

\item{\code{y_hat}}{matrix.}

\item{\code{e_hat}}{matrix.}

\item{\code{df_linreg}}{numeric.}

\item{\code{var_sigma_hat}}{numeric.}

\item{\code{p_value}}{matrix.}

\item{\code{t_value}}{matrix.}

\item{\code{var_beta_hat}}{matrix.}

\item{\code{data_name}}{character.}
}}


\examples{
linreg_mod <- linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
}
