\name{lab4new-package}
\alias{lab4new-package}
\alias{lab4new}
\docType{package}
\title{
\packageTitle{lab4new}
}
\description{
\packageDescription{lab4new}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab4new}
\packageIndices{lab4new}
Package contains of implementations of two algorithm. Dijkstra is for finding the shortest path from the initial node and Euclidean is for calculating greatest common divisior of two numbers.
}
\author{ packageAuthor {Farha}

Maintainer packageMaintainer{Farha}
}
\references{
https://en.wikipedia.org/wiki/Euclideanalgorithm.
https://en.wikipedia.org/wiki/Dijkstra%27salgorithm
}

\keyword{ package }
\seealso{
https://en.wikipedia.org/wiki/Greatest_common_divisor.
https://www.youtube.com/watch?v=gdmfOwyQlcI.
}
\examples{
linreg_mod <- linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)
linreg_mod$summary()
}


